Use per-project .bowerrc, instead of per-app.
If we create and use a .bowerrc in every app, then we lose
the ability to control where bower dependencies are installed.
This would be fine if bower were installing dependencies for
a language with a well-defined import model, but in reality
bower manages javascript and CSS resources.

Using a .bowerrc in the project root lets us control the
installation directory of dependencies, ensures that different
dependency versions are not installed in different directories,
and doesn't require sudo if for some reason a dependency needs
to be installed through the system package manager.